Moved functions for generating function calls to FunctionCalls
