+++ code/drasil-code/drasil-code.cabal
@@ -22,6 +22,7 @@ library
     , Language.Drasil.Code.Imperative.Helpers
     , Language.Drasil.Code.Imperative.Data
     , Language.Drasil.Code.Imperative.Descriptions
+    , Language.Drasil.Code.Imperative.FunctionCalls
     , Language.Drasil.Code.Imperative.GenerateGOOL
     , Language.Drasil.Code.Imperative.Generator
     , Language.Drasil.Code.Imperative.Import
@@ -77,6 +78,7 @@ executable codegenTest
     , Language.Drasil.Code.Imperative.Helpers
     , Language.Drasil.Code.Imperative.Data
     , Language.Drasil.Code.Imperative.Descriptions
+    , Language.Drasil.Code.Imperative.FunctionCalls
     , Language.Drasil.Code.Imperative.GenerateGOOL
     , Language.Drasil.Code.Imperative.Generator
     , Language.Drasil.Code.Imperative.Import
